requirements.txt
flask
selenium
playwright
pillow
beautifulsoup4
requests
css-html-js-minify